+++ api/src/main/java/com/api/api/patient/PatientController.java
@@ -1,13 +1,21 @@
 package com.api.api.patient;
 
+import com.api.api.token.TokenJWT;
+import com.api.api.token.TokenManager;
+import com.api.api.utils.roles;
+import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 
 @RestController
 @RequestMapping(path = "/patient")
 public class PatientController {
+
+    private static final String PATIENT_ROLE = roles.PATIENT.getValue();
 
     private final PatientService patientService;
 
@@ -18,6 +26,10 @@ public class PatientController {
 
     @PostMapping(path = "/create", consumes = "application/json")
     public ResponseEntity<String> createPatient(@RequestBody PatientDAO patient){
+
+        if(patient.isValidToCreate())
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"Wrong parameters.");
+
         PatientDAO patientDAO = patientService.createPatient(patient);
 
         return ResponseEntity.ok("Patient: " + patientDAO.getEmail() + " has been created.");
@@ -26,11 +38,24 @@ public class PatientController {
     @GetMapping(path = "/auth")
     public ResponseEntity<String> authPatient(PatientDAO patient){
 
-        return null;
+        if(!patient.isValidAuth())
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"Wrong parameters.");
+
+        return ResponseEntity.ok(patientService.auth(patient.getEmail(), patient.getPwd()));
     }
 
     @PutMapping(path = "/{email}/update", consumes = "application/json")
-    public ResponseEntity<String> updatePatient(@RequestBody PatientDAO patient){
+    public ResponseEntity<String> updatePatient(@RequestHeader("Authorization") String token,
+                                                @RequestBody PatientDAO patient,
+                                                @PathVariable String email){
+
+        if(!patient.isValidToUpdate())
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"Wrong parameters.");
+
+        TokenJWT tokenJWT = TokenManager.isTokenValid(token);
+        if(tokenJWT == null  || !tokenJWT.getEmail().equals(email))
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED,"Not allowed.");
+
 
         PatientDAO patientDAO = patientService.updatePatient(patient);
 
@@ -38,7 +63,12 @@ public class PatientController {
     }
 
     @DeleteMapping(path = "/{email}/delete")
-    public ResponseEntity<String> deletePatient(@PathVariable String email){
+    public ResponseEntity<String> deletePatient(@RequestHeader("Authorization") String token,
+                                                @PathVariable String email){
+
+        TokenJWT tokenJWT = TokenManager.isTokenValid(token);
+        if(tokenJWT == null  || !tokenJWT.getEmail().equals(email))
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED,"Not allowed.");
 
         String patientEmail = patientService.deletePatient(email);
 
@@ -46,7 +76,12 @@ public class PatientController {
     }
 
     @GetMapping(path = "/{email}/get", produces = "application/json")
-    public ResponseEntity<PatientDAO> getPatient(@PathVariable String email){
+    public ResponseEntity<PatientDAO> getPatient(@RequestHeader("Authorization") String token,
+                                                 @PathVariable String email){
+
+        TokenJWT tokenJWT = TokenManager.isTokenValid(token);
+        if(tokenJWT == null  || (!tokenJWT.getEmail().equals(email) && tokenJWT.getRole().equals(PATIENT_ROLE)))
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED,"Not allowed.");
 
         PatientDAO patientDAO = patientService.getPatient(email);
